user management

+++ user_management/views.py
@@ -9,6 +9,8 @@ from django.contrib import messages
 
 @login_required
 def dashboard_users(request):
+    if not request.user.is_superuser:  
+        return redirect('login')
     user_count = User.objects.count()
     group_count = Group.objects.count()
 
@@ -38,11 +40,16 @@ def manage_inactive_users(request):
     page_number = request.GET.get('page')
     page_obj = paginator.get_page(page_number)
 
+    for index, inactive_users in enumerate(page_obj.object_list):
+        inactive_users.serial_number = index + 1 + (page_obj.number - 1) * paginator.per_page
+
     return render(request, 'user_management/manage_inactive_users.html', {'page_obj': page_obj})
 
 
 @login_required
 def user_list(request):
+    if not request.user.is_superuser:  
+        return redirect('login')
     users = User.objects.all()
     paginator = Paginator(users, 10)  
     page_number = request.GET.get('page')
@@ -51,6 +58,8 @@ def user_list(request):
 
 @login_required
 def user_add(request):
+    if not request.user.is_superuser:  
+        return redirect('login')
     if request.method == 'POST':
         form = UserCreationForm(request.POST)
         if form.is_valid():
@@ -62,6 +71,8 @@ def user_add(request):
 
 @login_required
 def user_edit(request, pk):
+    if not request.user.is_superuser:  
+        return redirect('login')
     user = get_object_or_404(User, pk=pk)
     if request.method == 'POST':
         form = UserChangeForm(request.POST, instance=user)
@@ -74,17 +85,23 @@ def user_edit(request, pk):
 
 @login_required
 def user_delete(request, pk):
+    if not request.user.is_superuser:  
+        return redirect('login')
     user = get_object_or_404(User, pk=pk)
     user.delete()
     return redirect('user_list')
 
 @login_required
 def group_list(request):
+    if not request.user.is_superuser:  
+        return redirect('login')
     groups = Group.objects.all()
     return render(request, 'user_management/group_list.html', {'groups': groups})
 
 @login_required
 def group_add(request):
+    if not request.user.is_superuser:  
+        return redirect('login')
     if request.method == 'POST':
         name = request.POST.get('name')
         Group.objects.create(name=name)
@@ -92,6 +109,8 @@ def group_add(request):
 
 @login_required
 def group_edit(request, pk):
+    if not request.user.is_superuser:  
+        return redirect('login')
     group = get_object_or_404(Group, pk=pk)
     if request.method == 'POST':
         group.name = request.POST.get('name')
@@ -100,6 +119,8 @@ def group_edit(request, pk):
 
 @login_required
 def group_delete(request, pk):
+    if not request.user.is_superuser:  
+        return redirect('login')
     group = get_object_or_404(Group, pk=pk)
     group.delete()
     return redirect('group_list')
@@ -108,6 +129,8 @@ def group_delete(request, pk):
 # Ministry Views
 @login_required
 def ministry_list_create(request):
+    if not request.user.is_superuser:  
+        return redirect('login')
     if request.method == 'POST':  
         name = request.POST['name']
         description = request.POST.get('description', '')
@@ -129,6 +152,8 @@ def ministry_list_create(request):
 
 @login_required
 def ministry_edit(request, pk):
+    if not request.user.is_superuser:  
+        return redirect('login')
     ministry = get_object_or_404(Ministry, pk=pk)
     if request.method == 'POST':  
         ministry.name = request.POST['name']
@@ -140,6 +165,8 @@ def ministry_edit(request, pk):
     return redirect('ministry_list_create')
 
 def ministry_delete(request, pk):
+    if not request.user.is_superuser:  
+        return redirect('login')
     ministry = get_object_or_404(Ministry, pk=pk)
     ministry.delete()
     messages.success(request, f'{ministry.name}, have been deleted successfully')
@@ -148,6 +175,8 @@ def ministry_delete(request, pk):
 
 # Department Views
 def department_list_create(request):
+    if not request.user.is_superuser:  
+        return redirect('login')
     
     if request.method == 'POST':  
         name = request.POST['name']
@@ -175,6 +204,8 @@ def department_list_create(request):
 
 
 def department_edit(request, pk):
+    if not request.user.is_superuser:  
+        return redirect('login')
     department = get_object_or_404(Department, pk=pk)
     ministries = Ministry.objects.all()
 
@@ -195,6 +226,8 @@ def department_edit(request, pk):
 
 
 def department_delete(request, pk):
+    if not request.user.is_superuser:  
+        return redirect('login')
     department = get_object_or_404(Department, pk=pk)
     department.delete()
     messages.success(request, f'{department.name}, have been deleted successfully')
